refactor: Rewrite Assignment as ValueAssignment
Making place for `FunctionAssignment` in the next PR.
